--- src/05_hash_function.ts
@@ -1,60 +1,0 @@
-import { sha256 } from "js-sha256";
-import { serialize } from "serializer.ts/Serializer";
-
-export type Address = string;
-
-export class Transaction {
-  public senderAddress: Address;
-  public recipientAddress: Address;
-  public value: number;
-
-  constructor(senderAddress: Address, recipientAddress: Address, value: number) {
-    this.senderAddress = senderAddress;
-    this.recipientAddress = recipientAddress;
-    this.value = value;
-  }
-}
-
-export class Block {
-  public blockNumber: number;
-  public transactions: Array<Transaction>;
-  public timestamp: number;
-  public nonce: number;
-  public prevBlock: string;
-
-  constructor(blockNumber: number, transactions: Array<Transaction>, timestamp: number, nonce: number,
-    prevBlock: string) {
-    this.blockNumber = blockNumber;
-    this.transactions = transactions;
-    this.timestamp = timestamp;
-    this.nonce = nonce;
-    this.prevBlock = prevBlock;
-  }
-
-  // Calculates the SHA256 of the entire block, including its transactions.
-  public sha256(): string {
-    return sha256(JSON.stringify(serialize<Block>(this)));
-  }
-}
-
-export class Blockchain {
-  public nodeId: string;
-  public blocks: Array<Block>;
-  public transactionPool: Array<Transaction>;
-
-  constructor(nodeId: string) {
-    this.nodeId = nodeId;
-    this.blocks = [];
-    this.transactionPool = [];
-  }
-
-  // Submits new transaction
-  public submitTransaction(senderAddress: Address, recipientAddress: Address, value: number) {
-    this.transactionPool.push(new Transaction(senderAddress, recipientAddress, value));
-  }
-
-  // Creates new block on the blockchain.
-  public createBlock() {
-    // TBD
-  }
-}
