--- src/07_add_new_block.ts
@@ -1,129 +1,0 @@
-import { sha256 } from "js-sha256";
-import { serialize } from "serializer.ts/Serializer";
-import BigNumber from "bignumber.js";
-
-export type Address = string;
-
-export class Transaction {
-  public senderAddress: Address;
-  public recipientAddress: Address;
-  public value: number;
-
-  constructor(senderAddress: Address, recipientAddress: Address, value: number) {
-    this.senderAddress = senderAddress;
-    this.recipientAddress = recipientAddress;
-    this.value = value;
-  }
-}
-
-export class Block {
-  public blockNumber: number;
-  public transactions: Array<Transaction>;
-  public timestamp: number;
-  public nonce: number;
-  public prevBlock: string;
-
-  constructor(blockNumber: number, transactions: Array<Transaction>, timestamp: number, nonce: number,
-    prevBlock: string) {
-    this.blockNumber = blockNumber;
-    this.transactions = transactions;
-    this.timestamp = timestamp;
-    this.nonce = nonce;
-    this.prevBlock = prevBlock;
-  }
-
-  // Calculates the SHA256 of the entire block, including its transactions.
-  public sha256(): string {
-    return sha256(JSON.stringify(serialize<Block>(this)));
-  }
-}
-
-export class Blockchain {
-  // Let's define that our "genesis" block as an empty block, starting from now.
-  public static readonly GENESIS_BLOCK = new Block(0, [], Blockchain.now(), 0, "");
-
-  public static readonly DIFFICULTY = 4;
-  public static readonly TARGET = 2 ** (256 - Blockchain.DIFFICULTY);
-
-  public nodeId: string;
-  public blocks: Array<Block>;
-  public transactionPool: Array<Transaction>;
-
-  constructor(nodeId: string) {
-    this.nodeId = nodeId;
-    this.blocks = [Blockchain.GENESIS_BLOCK];
-    this.transactionPool = [];
-  }
-
-  // Validates PoW.
-  public static isPoWValid(pow: string): boolean {
-    try {
-      if (!pow.startsWith("0x")) {
-        pow = `0x${pow}`;
-      }
-
-      return new BigNumber(pow).lessThanOrEqualTo(Blockchain.TARGET.toString());
-    } catch {
-      return false;
-    }
-  }
-
-  // Mines for block.
-  public mineBlock(transactions: Array<Transaction>): Block {
-    // Create a new block which will "point" to the last block.
-    const lastBlock = this.getLastBlock();
-    const newBlock = new Block(lastBlock.blockNumber + 1, transactions, Blockchain.now(), 0, lastBlock.sha256());
-
-    while (true) {
-      const pow = newBlock.sha256();
-      console.log(`Mining #${newBlock.blockNumber}: nonce: ${newBlock.nonce}, pow: ${pow}`);
-
-      if (Blockchain.isPoWValid(pow)) {
-        console.log(`Found valid POW: ${pow}!`);
-        break;
-      }
-
-      newBlock.nonce++;
-    }
-
-    return newBlock;
-  }
-
-  // Submits new transaction
-  public submitTransaction(senderAddress: Address, recipientAddress: Address, value: number) {
-    this.transactionPool.push(new Transaction(senderAddress, recipientAddress, value));
-  }
-
-  // Creates new block on the blockchain.
-  public createBlock(): Block {
-    // Mine the transactions in a new block.
-    const newBlock = this.mineBlock(this.transactionPool);
-
-    // Append the new block to the blockchain.
-    this.blocks.push(newBlock);
-
-    // Remove the mined transactions.
-    this.transactionPool = [];
-
-    return newBlock;
-  }
-
-  public getLastBlock(): Block {
-    return this.blocks[this.blocks.length - 1];
-  }
-
-  public static now(): number {
-    return Math.round(new Date().getTime() / 1000);
-  }
-}
-
-const blockchain = new Blockchain("node123");
-
-blockchain.submitTransaction("Alice", "Bob", 1000);
-blockchain.submitTransaction("Alice", "Eve", 12345);
-
-const block = blockchain.createBlock();
-
-console.log(`Mined block: ${JSON.stringify(serialize(block))}`);
-console.log(`Mined block with: ${block.sha256()}`);
-
