+++ src/address.ts
@@ -1,0 +1,3 @@
+type Address = string;
+
+export default Address
