+++ src/blockchain.ts
@@ -1,71 +1,15 @@
-import { sha256 } from "js-sha256";
-import { serialize, deserialize } from "serializer.ts/Serializer";
-import BigNumber from "bignumber.js";
-
+import { Set } from "typescript-collections";
+import deepEqual = require("deep-equal");
 import * as fs from "fs";
 import * as path from "path";
-import deepEqual = require("deep-equal");
+import { serialize, deserialize } from "serializer.ts/Serializer";
+import Block from './block'
+import Transaction from './transaction'
+import Node from './node'
+import Address from './address'
+import BigNumber from "bignumber.js";
 
-import * as uuidv4 from "uuid/v4";
-import * as express from "express";
-import * as bodyParser from "body-parser";
-import { URL } from "url";
-import axios from "axios";
-
-import { Set } from "typescript-collections";
-import * as parseArgs from "minimist";
-
-export type Address = string;
-
-export class Transaction {
-  public senderAddress: Address;
-  public recipientAddress: Address;
-  public value: number;
-
-  constructor(senderAddress: Address, recipientAddress: Address, value: number) {
-    this.senderAddress = senderAddress;
-    this.recipientAddress = recipientAddress;
-    this.value = value;
-  }
-}
-
-export class Block {
-  public blockNumber: number;
-  public transactions: Array<Transaction>;
-  public timestamp: number;
-  public nonce: number;
-  public prevBlock: string;
-
-  constructor(blockNumber: number, transactions: Array<Transaction>, timestamp: number, nonce: number,
-    prevBlock: string) {
-    this.blockNumber = blockNumber;
-    this.transactions = transactions;
-    this.timestamp = timestamp;
-    this.nonce = nonce;
-    this.prevBlock = prevBlock;
-  }
-
-  // Calculates the SHA256 of the entire block, including its transactions.
-  public sha256(): string {
-    return sha256(JSON.stringify(serialize<Block>(this)));
-  }
-}
-
-export class Node {
-  public id: string;
-  public url: URL;
-
-  constructor(id: string, url: URL) {
-    this.id = id;
-    this.url = url;
-  }
-
-  public toString(): string {
-      return `${this.id}:${this.url}`;
-  }
-}
-
-export class Blockchain {
+export default class Blockchain {
   // Let's define that our "genesis" block as an empty block, starting from the January 1, 1970 (midnight "UTC").
   public static readonly GENESIS_BLOCK = new Block(0, [], 0, 0, "fiat lux");
 
@@ -268,131 +212,4 @@ export class Blockchain {
   public static now(): number {
     return Math.round(new Date().getTime() / 1000);
   }
-}
-
-// Web server:
-const ARGS = parseArgs(process.argv.slice(2));
-const PORT = ARGS.port || 3000;
-const app = express();
-const nodeId = ARGS.id || uuidv4();
-const blockchain = new Blockchain(nodeId);
-
-// Set up bodyParser:
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(bodyParser.json());
-app.use((err: any, req: express.Request, res: express.Response, next: express.NextFunction) => {
-  console.error(err.stack);
-
-  res.status(500);
-});
-
-// Show all the blocks.
-app.get("/blocks", (req: express.Request, res: express.Response) => {
-  res.json(serialize(blockchain.blocks));
-});
-
-// Show specific block.
-app.get("/blocks/:id", (req: express.Request, res: express.Response) => {
-  const id = Number(req.params.id);
-  if (isNaN(id)) {
-    res.json("Invalid parameter!");
-    res.status(500);
-    return;
-  }
-
-  if (id >= blockchain.blocks.length) {
-    res.json(`Block #${id} wasn't found!`);
-    res.status(404);
-    return;
-  }
-
-  res.json(serialize(blockchain.blocks[id]));
-});
-
-app.post("/blocks/mine", (req: express.Request, res: express.Response) => {
-  // Mine the new block.
-  const newBlock = blockchain.createBlock();
-
-  res.json(`Mined new block #${newBlock.blockNumber}`);
-});
-
-// Show all transactions in the transaction pool.
-app.get("/transactions", (req: express.Request, res: express.Response) => {
-  res.json(serialize(blockchain.transactionPool));
-});
-
-app.post("/transactions", (req: express.Request, res: express.Response) => {
-  const senderAddress = req.body.senderAddress;
-  const recipientAddress = req.body.recipientAddress;
-  const value = Number(req.body.value);
-
-  if (!senderAddress || !recipientAddress || !value)  {
-    res.json("Invalid parameters!");
-    res.status(500);
-    return;
-  }
-
-  blockchain.submitTransaction(senderAddress, recipientAddress, value);
-
-  res.json(`Transaction from ${senderAddress} to ${recipientAddress} was added successfully`);
-});
-
-app.get("/nodes", (req: express.Request, res: express.Response) => {
-  res.json(serialize(blockchain.nodes.toArray()));
-});
-
-app.post("/nodes", (req: express.Request, res: express.Response) => {
-  const id = req.body.id;
-  const url = new URL(req.body.url);
-
-  if (!id || !url)  {
-    res.json("Invalid parameters!");
-    res.status(500);
-    return;
-  }
-
-  const node = new Node(id, url);
-
-  if (blockchain.register(node)) {
-    res.json(`Registered node: ${node}`);
-  } else {
-    res.json(`Node ${node} already exists!`);
-    res.status(500);
-  }
-});
-
-app.put("/nodes/consensus", (req: express.Request, res: express.Response) => {
-  // Fetch the state of the other nodes.
-  const requests = blockchain.nodes.toArray().map(node => axios.get(`${node.url}blocks`));
-
-  if (requests.length === 0) {
-    res.json("There are nodes to sync with!");
-    res.status(404);
-
-    return;
-  }
-
-  axios.all(requests).then(axios.spread((...blockchains) => {
-    if (blockchain.consensus(blockchains.map(res => deserialize<Block[]>(Block, res.data)))) {
-      res.json(`Node ${nodeId} has reached a consensus on a new state.`);
-    } else {
-      res.json(`Node ${nodeId} hasn't reached a consensus on the existing state.`);
-    }
-
-    res.status(200);
-    return;
-  })).catch(err => {
-    console.log(err);
-    res.status(500);
-    res.json(err);
-    return;
-  });
-
-  res.status(500);
-});
-
-if (!module.parent) {
-  app.listen(PORT);
-
-  console.log(`Web server started on port ${PORT}. Node ID is: ${nodeId}`);
 }
